Move RemoveFilesDialogFragment to another package
