+++ owncloudApp/src/main/java/com/owncloud/android/ui/preview/PreviewImageFragment.kt
@@ -50,7 +50,7 @@ import com.owncloud.android.domain.files.model.OCFile
 import com.owncloud.android.files.FileMenuFilter
 import com.owncloud.android.ui.controller.TransferProgressController
 import com.owncloud.android.ui.dialog.ConfirmationDialogFragment
-import com.owncloud.android.ui.dialog.RemoveFilesDialogFragment
+import com.owncloud.android.presentation.ui.files.removefile.RemoveFilesDialogFragment
 import com.owncloud.android.ui.fragment.FileFragment
 import com.owncloud.android.utils.PreferenceUtils
 import timber.log.Timber
